pull out table generation
